MavLinkDrone: Fixes #443 and makes sysid/compid use the expected short type
sysid/compid need to use short with the current library and not byte as it stores the values as an unsigned 0-255 value in the short variable. Missing the conversion will cause a crash.

+++ ClientLib/src/main/java/org/droidplanner/services/android/impl/core/drone/autopilot/MavLinkDrone.java
@@ -27,9 +27,9 @@ public interface MavLinkDrone extends Drone {
 
     void onMavLinkMessageReceived(MAVLinkMessage message);
 
-    public byte getSysid();
+    public short getSysid();
 
-    public byte getCompid();
+    public short getCompid();
 
     public State getState();
